+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 - `calculations`: Removed `getRadian` export
 - `data`: Removed `setField` export, use [`lodash.set(object, path, value)`](https://lodash.com/docs/4.17.15#set) instead
+- `data`: Removed `gatherArrays` export
 - `formatters`: Remove `capitalizeFirst` export, use [`lodash.upperFirst(string)`](https://lodash.com/docs/4.17.15#upperFirst) instead